Add a new tip and Fetch tips
Places will be a second branch of the Firebase database, and tips will
be organized by unique placeID. Added a Tip model in Tip.swift and methods for
adding a tip and fetching an array of a place's tips in DatabaseRequests.

Added placeholder buttons to nav bar for adding a tip and fetching tips
in order to test AddTip and FetchTips functions. Verified both are ready
for populating the detailVC.